Add a logging service
The intention of this change is to enable logging functionality into the
application so that it can be properly instrumented.

Replace the weatherData service with a WeatherService service

That title isn't the most indicative, I'll admit, but this change
replaces a service that returned a fixed data payload with a service
that can return data in a variety of ways and permutations, making it
much more compact, yet providing much more functionality to the
application.