Update the Docker PHP service configuration
This change upgrades the PHP service to use PHP 8.0, and also refactors
the service to be built from a Dockerfile so that all of the required
runtime dependencies are installed, enabled, and available. These
include PHP's intl extension, xdebug for debugging, Composer (along with
its dependencies of Git and the Zip extension).