Refactor the UX to be more fully responsive
This change refactors the design so that the UX works at as many breakpoints as
possible. Specifically, it does this by adjusting the widths of various
elements at various breakpoints, so that they the respective contents
are more readable, and also adjust some of the items to conditionally
render in a shorter or longer version, which make sense at that
breakpoint.