Add laminas-validator, Codeception, and Pest PHP as dependencies
While I should have added these as separate commits, it's not possible
at the moment, so I'm lumping them in together.

laminas-validator is required to validate the start and/or end date, if
they're supplied to a controller.

Codeception isn't required, at least not now, but may be later, as I
usually use it for integration testing. Pest PHP however, is going to be
the primary testing framework in this application, mainly because I want
to check it out and give it a meaningful trial.