Refactor the default route's logic into a separate class
This change has been coming for a while now, as I've increasingly found
the bootstrap file (public/index.php) to be becoming rather bloated and
hard to maintain. Given that, I refactored the logic of the search
results route's handler into a separate class. This has the benefit of
reducing the size of the bootstrap file and making the functionality
easier to test (okay, makes it properly testable).