Add PHP CS Support
This change adds support for PHP Code Style/Sniffer to help improve the
quality of the code in the codebase. It also adds a configuration file
for it, one copied from Laminas/Mezzio projects. To complete that, it
also adds Laminas Coding Style, as that's the style that I use.